lambdas/requirements.txt
requests==2.31.0
psycopg2-binary==2.9.9
openai==1.12.0
anthropic==0.18.1
boto3==1.34.0
python-dotenv==1.0.1
